[#2597] Replaced in-repo persistent_deploy module with 'drupal/deploy_steps' contrib.
