+++ .vortex/tooling/src/provision
@@ -406,12 +406,12 @@ drush cache:rebuild
 pass "Cache was rebuilt."
 echo
 
-# Enable the persistent deploy runner and the always-on site modules before
-# deploy hooks, so their deploy step plugins and run-once deploy hooks run on
-# this provision. Site module names are project-specific (the installer renames
+# Enable the deploy_steps runner and the always-on site modules before deploy
+# hooks, so their deploy step plugins and run-once deploy hooks run on this
+# provision. Site module names are project-specific (the installer renames
 # them), so resolve them by glob. Idempotent in every environment.
 task "Enabling site modules."
-drush pm:install persistent_deploy
+drush pm:install deploy_steps
 for module_path in "${WEBROOT}"/modules/custom/*_base "${WEBROOT}"/modules/custom/*_migrate; do
   if [ -d "${module_path}" ]; then
     drush pm:install "$(basename "${module_path}")"
