+++ composer.json
@@ -15,6 +15,7 @@
         "drupal/config_update": "^2@alpha",
         "drupal/core-composer-scaffold": "~11.3.11",
         "drupal/core-recommended": "~11.3.11",
+        "drupal/deploy_steps": "^1.0.0-alpha1",
         "drupal/drupal_helpers": "^2.0",
         "drupal/environment_indicator": "^4.0.25",
         "drupal/generated_content": "^2.0",
