+++ scripts/README.md
@@ -2,6 +2,6 @@ This directory contains Vortex and your custom scripts.
 
 Add your custom per-project scripts into the `./scripts` directory.
 
-Run-on-every-deploy logic belongs in a `PersistentDeploy` plugin instead - see `docs/deployment.md`.
+Run-on-every-deploy logic belongs in a `DeployStep` plugin instead - see `docs/deployment.md`.
 
 Do not modify scripts shipped by Vortex.
